refactor: #237 Update User Space Invite UX Flow
