+++ src/app/space/space.const.ts
@@ -1,0 +1,1 @@
+export const COOKIE_VALUE_SAVE_SPACE_ID_PHASE = "saveSpaceIdPhase";
