+++ src/component/login/index.ts
@@ -1,1 +1,2 @@
 export { SocialLoginButton } from "./SocialLoginButton";
+export { SocialLoginArea } from "./SocialLoginArea";
