+++ src/hooks/api/login/usePostSignUp.ts
@@ -4,7 +4,9 @@ import Cookies from "js-cookie";
 import { useNavigate } from "react-router-dom";
 
 import { api } from "@/api";
+import { COOKIE_VALUE_SAVE_SPACE_ID_PHASE } from "@/app/space/space.const.ts";
 import { PATHS } from "@/config/paths";
+import { useApiJoinSpace } from "@/hooks/api/space/useApiJoinSpace.ts";
 import { useToast } from "@/hooks/useToast";
 import { authAtom } from "@/store/auth/authAtom";
 import { AuthResponse, SocialLoginKind } from "@/types/loginType";
@@ -19,6 +21,7 @@ export const usePostSignUp = () => {
   const navigate = useNavigate();
   const { toast } = useToast();
   const [, setAuth] = useAtom(authAtom);
+  const { mutate } = useApiJoinSpace();
   const signUpWithToken = async ({ accessToken, name, socialType }: PostSignUp): Promise<AuthResponse> => {
     const response = await api.post<AuthResponse>(
       "/api/auth/sign-up",
@@ -44,8 +47,20 @@ export const usePostSignUp = () => {
         Cookies.set("refreshToken", data.refreshToken, { expires: 7 });
         setAuth({ isLogin: true, name: data.name, email: data.email, memberRole: data.memberRole, imageUrl: data.imageUrl });
       }
-      toast.success("Layer에 온걸 환영해요!");
-      navigate(PATHS.home());
+
+      const saveSpaceIdPhase = Cookies.get(COOKIE_VALUE_SAVE_SPACE_ID_PHASE);
+      if (saveSpaceIdPhase) {
+        mutate(parseInt(saveSpaceIdPhase), {
+          onSuccess: () => {
+            toast.success("첫 스페이스에 오신 걸 환영해요!");
+            navigate(PATHS.spaceDetail(saveSpaceIdPhase));
+            Cookies.remove(COOKIE_VALUE_SAVE_SPACE_ID_PHASE);
+          },
+        });
+      } else {
+        toast.success("Layer에 오신 걸 환영해요!");
+        navigate(PATHS.home());
+      }
     },
     onError: (error) => {
       console.error("Sign up failed:", error);
