+++ src/hooks/api/space/useApiJoinSpace.ts
@@ -1,12 +1,13 @@
 import { useMutation } from "@tanstack/react-query";
-import { useNavigate } from "react-router-dom";
 
 import { api } from "@/api";
-import { useToast } from "@/hooks/useToast";
+
+type ErrorType = {
+  status: number;
+  message: string;
+};
 
 export const useApiJoinSpace = () => {
-  const navigate = useNavigate();
-  const { toast } = useToast();
   const joinSpace = async (spaceId: number) => {
     const res = await api.post(`/api/space/join`, null, { params: { spaceId } });
     return res;
@@ -14,14 +15,6 @@ export const useApiJoinSpace = () => {
 
   return useMutation({
     mutationFn: (spaceId: number) => joinSpace(spaceId),
-    onSuccess: () => {
-      toast.success("팀에 초대 되었습니다.");
-      navigate("/");
-    },
-    onError: (error: { status: number }) => {
-      if (error.status === 400) {
-        toast.error("이미 참여한 스페이스 입니다");
-      }
-    },
+    onError: (error: ErrorType) => error,
   });
 };
